Adding tests for App WIP make VMs app agnostic

+++ ModernKeePass/ViewModels/MainVm.cs
@@ -49,14 +49,14 @@ namespace ModernKeePass.ViewModels
         {
             var app = (App)Application.Current;
             var mru = StorageApplicationPermissions.MostRecentlyUsedList;
-            var isDatabaseOpen = app.Database != null && app.Database.Status == DatabaseHelper.DatabaseStatus.Opened;
+            var isDatabaseOpen = app.Database != null && app.Database.Status == (int) DatabaseHelper.DatabaseStatus.Opened;
 
             var mainMenuItems = new ObservableCollection<MainMenuItemVm>
             {
                 new MainMenuItemVm
                 {
                     Title = "Open", PageType = typeof(OpenDatabasePage), Destination = destinationFrame, Parameter = referenceFrame, SymbolIcon = Symbol.Page2,
-                    IsSelected = app.Database.Status == DatabaseHelper.DatabaseStatus.Opening
+                    IsSelected = app.Database.Status == (int) DatabaseHelper.DatabaseStatus.Opening
                 },
                 new MainMenuItemVm
                 {
@@ -69,7 +69,7 @@ namespace ModernKeePass.ViewModels
                 },
                 new MainMenuItemVm {
                     Title = "Recent" , PageType = typeof(RecentDatabasesPage), Destination = destinationFrame, Parameter = referenceFrame, SymbolIcon = Symbol.Copy,
-                    IsSelected = (app.Database == null || app.Database.Status == DatabaseHelper.DatabaseStatus.Closed) && mru.Entries.Count > 0, IsEnabled = mru.Entries.Count > 0
+                    IsSelected = (app.Database == null || app.Database.Status == (int) DatabaseHelper.DatabaseStatus.Closed) && mru.Entries.Count > 0, IsEnabled = mru.Entries.Count > 0
                 },
                 new MainMenuItemVm
                 {
@@ -78,7 +78,7 @@ namespace ModernKeePass.ViewModels
             };
             // Auto-select the Recent Items menu item if the conditions are met
             SelectedItem = mainMenuItems.FirstOrDefault(m => m.IsSelected);
-            if (app.Database != null && app.Database.Status == DatabaseHelper.DatabaseStatus.Opened)
+            if (app.Database != null && app.Database.Status == (int) DatabaseHelper.DatabaseStatus.Opened)
                 mainMenuItems.Add(new MainMenuItemVm
                 {
                     Title = app.Database.Name,
